locust-testing/pyproject.toml
[project]
name = "locust-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "locust>=2.37.10",
    "python-dotenv>=1.1.0",
]
